requirements.txt
bottle
bottle-sqlalchemy
psycopg2
